% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCreate.R
\name{create_test_files}
\alias{create_test_files}
\title{Create Test Files}
\usage{
create_test_files(package_dir = getwd(), file_per_function = TRUE,
  dbg = TRUE)
}
\arguments{
\item{package_dir}{path to package directory in which to create the test
files}

\item{file_per_function}{if \code{TRUE} (default), one test file
\code{test-<function>.R} is generated for each function, otherwise one test
file \code{test-<source-file>} is generated for each source file.}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Create test files for each source file containing one
  \code{\link[testthat]{test_that}} call for each function in the package
}
\details{
Existing test files will not be overwritten.
}
