% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMain.R
\name{printTestMessage}
\alias{printTestMessage}
\title{Print a Test with its Result}
\usage{
printTestMessage(testexpression, testresult, newline = TRUE)
}
\arguments{
\item{testexpression}{text description of what was tested}

\item{testresult}{boolean result (of length one) of the test}

\item{newline}{if \code{TRUE} (default) a new line character is appended
to the message shown.}
}
\value{
the message that was shown as a character string
}
\description{
Print a test with its result as a message and return the message as a
character string
}
\examples{
printTestMessage("apple == apple", 1 == 1)
printTestMessage("apple == pear", 1 == 2)
}
