% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMain.R
\name{saveArgs}
\alias{saveArgs}
\title{Save the Arguments and Result of a Function Call}
\usage{
saveArgs(functionName, ...,
  targetdir = kwb.utils::createDirectory(file.path(tempdir(), "test")))
}
\arguments{
\item{functionName}{name of the function to which the arguments to be saved
belong. It will be used to generate a file name for the RData file.}

\item{...}{named arguments representing the arguments that have been given
to the function \code{functionName}.}

\item{targetdir}{directory in which to store the objects given in \code{...}
Default: subdirectory \code{test} in \code{tempdir()}}
}
\value{
path to the file written (invisibly)
}
\description{
Save the list of named arguments given in \code{...} to an RData file
\code{args_<functionName>_<hhmmss>_<no>.RData} in the directory given in
\code{targetdir}. This function can be used to log the inputs given to a
function together with the result returned by the function.
\code{\link{test_function}} can then be used to check whether another version
of the function (e.g. obtained by code cleaning) can reproduce the stored
results from the stored arguments. Check out the example on the help page for
\code{\link{test_function}}.
}
\seealso{
\code{\link{loadArgs}}
}
