% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMain.R
\name{testColumnwiseIdentity}
\alias{testColumnwiseIdentity}
\title{Check Corresponding Columns in two Data Frames for Identity}
\usage{
testColumnwiseIdentity(...)
}
\arguments{
\item{...}{two data frames given as named arguments. The argument names will
appear in the output. By doing so you can give a longer expression that
returns a data frame a short name 'on-the-fly'.}
}
\description{
For all columns in the first data frame, check if the second data frame
has identical values in columns of the same name
}
\examples{
# Compare two identical data frames. Give them short names data.1 and data.2
testColumnwiseIdentity(data.1 = (x <- data.frame(a = 1:2, b = 2:3)),
                       data.2 = x)

# Compare two data frames differing in one column
testColumnwiseIdentity(A = data.frame(x = 1:2, y = 2:3),
                       B = data.frame(x = 1:2, y = 3:4))
}

